improve: migrate decrypter modules to TypeScript
