+++ src/sources/jiosaavn.js
@@ -1,6 +1,6 @@
 import { PassThrough } from 'node:stream'
 import { encodeTrack, http1makeRequest, logger, getBestMatch, } from '../utils.ts'
-import { desEcbDecryptBase64ToUtf8 } from '../decrypters/des-ecb.js'
+import { desEcbDecryptBase64ToUtf8 } from '../decrypters/des-ecb.ts'
 
 const API_BASE = 'https://www.jiosaavn.com/api.php'
 const J_BUFFER = Buffer.from('38346591')
